Fixed windtunnel range.py so it works with python3 and its int/int division converting to float...

+++ windtunnel/range.py
@@ -269,7 +269,7 @@ class Range(wx.Frame):
         #setup latitude and longitude grid
         res=1080
         nph=res
-        nth=res/2
+        nth=int(res/2)
 
         lon=np.zeros(nph)
         lat=np.zeros(nth)
